+++ client/tx/legacy.go
@@ -6,28 +6,28 @@ import (
 	"github.com/cosmos/cosmos-sdk/client"
 	"github.com/cosmos/cosmos-sdk/codec"
 	sdk "github.com/cosmos/cosmos-sdk/types"
+	"github.com/cosmos/cosmos-sdk/x/auth/legacy/legacytx"
 	"github.com/cosmos/cosmos-sdk/x/auth/signing"
-	"github.com/cosmos/cosmos-sdk/x/auth/types"
 )
 
 // ConvertTxToStdTx converts a transaction to the legacy StdTx format
-func ConvertTxToStdTx(codec *codec.LegacyAmino, tx signing.Tx) (types.StdTx, error) {
-	if stdTx, ok := tx.(types.StdTx); ok {
+func ConvertTxToStdTx(codec *codec.LegacyAmino, tx signing.Tx) (legacytx.StdTx, error) {
+	if stdTx, ok := tx.(legacytx.StdTx); ok {
 		return stdTx, nil
 	}
 
-	aminoTxConfig := types.StdTxConfig{Cdc: codec}
+	aminoTxConfig := legacytx.StdTxConfig{Cdc: codec}
 	builder := aminoTxConfig.NewTxBuilder()
 
 	err := CopyTx(tx, builder)
 	if err != nil {
 
-		return types.StdTx{}, err
+		return legacytx.StdTx{}, err
 	}
 
-	stdTx, ok := builder.GetTx().(types.StdTx)
+	stdTx, ok := builder.GetTx().(legacytx.StdTx)
 	if !ok {
-		return types.StdTx{}, fmt.Errorf("expected %T, got %+v", types.StdTx{}, builder.GetTx())
+		return legacytx.StdTx{}, fmt.Errorf("expected %T, got %+v", legacytx.StdTx{}, builder.GetTx())
 	}
 
 	return stdTx, nil
@@ -58,13 +58,13 @@ func CopyTx(tx signing.Tx, builder client.TxBuilder) error {
 	return nil
 }
 
-func ConvertAndEncodeStdTx(txConfig client.TxConfig, stdTx types.StdTx) ([]byte, error) {
+func ConvertAndEncodeStdTx(txConfig client.TxConfig, stdTx legacytx.StdTx) ([]byte, error) {
 	builder := txConfig.NewTxBuilder()
 
 	var theTx sdk.Tx
 
 	// check if we need a StdTx anyway, in that case don't copy
-	if _, ok := builder.GetTx().(types.StdTx); ok {
+	if _, ok := builder.GetTx().(legacytx.StdTx); ok {
 		theTx = stdTx
 	} else {
 		err := CopyTx(stdTx, builder)
