+++ x/auth/legacy/legacytx/amino_signing_test.go
@@ -1,4 +1,4 @@
-package types
+package legacytx
 
 import (
 	"testing"
