+++ x/auth/legacy/legacytx/codec.go
@@ -1,0 +1,10 @@
+package legacytx
+
+import (
+	"github.com/cosmos/cosmos-sdk/codec"
+)
+
+func init() {
+	var amino = codec.NewLegacyAmino()
+	amino.RegisterConcrete(StdTx{}, "cosmos-sdk/StdTx", nil)
+}
