--- x/auth/types/stdtx.go
@@ -1,501 +1,0 @@
-package types
-
-import (
-	"encoding/json"
-	"fmt"
-
-	"github.com/cosmos/cosmos-sdk/types/tx/signing"
-
-	"github.com/tendermint/tendermint/crypto"
-	yaml "gopkg.in/yaml.v2"
-
-	"github.com/cosmos/cosmos-sdk/codec"
-	"github.com/cosmos/cosmos-sdk/codec/legacy"
-	codectypes "github.com/cosmos/cosmos-sdk/codec/types"
-	"github.com/cosmos/cosmos-sdk/crypto/types/multisig"
-	sdk "github.com/cosmos/cosmos-sdk/types"
-	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
-)
-
-// MaxGasWanted defines the max gas allowed.
-const MaxGasWanted = uint64((1 << 63) - 1)
-
-// Deprecated: StdFee includes the amount of coins paid in fees and the maximum
-// gas to be used by the transaction. The ratio yields an effective "gasprice",
-// which must be above some miminum to be accepted into the mempool.
-type StdFee struct {
-	Amount sdk.Coins `json:"amount" yaml:"amount"`
-	Gas    uint64    `json:"gas" yaml:"gas"`
-}
-
-// Deprecated: NewStdFee returns a new instance of StdFee
-func NewStdFee(gas uint64, amount sdk.Coins) StdFee {
-	return StdFee{
-		Amount: amount,
-		Gas:    gas,
-	}
-}
-
-// GetGas returns the fee's (wanted) gas.
-func (fee StdFee) GetGas() uint64 {
-	return fee.Gas
-}
-
-// GetAmount returns the fee's amount.
-func (fee StdFee) GetAmount() sdk.Coins {
-	return fee.Amount
-}
-
-// Bytes returns the encoded bytes of a StdFee.
-func (fee StdFee) Bytes() []byte {
-	if len(fee.Amount) == 0 {
-		fee.Amount = sdk.NewCoins()
-	}
-
-	bz, err := legacy.Cdc.MarshalJSON(fee)
-	if err != nil {
-		panic(err)
-	}
-
-	return bz
-}
-
-// GasPrices returns the gas prices for a StdFee.
-//
-// NOTE: The gas prices returned are not the true gas prices that were
-// originally part of the submitted transaction because the fee is computed
-// as fee = ceil(gasWanted * gasPrices).
-func (fee StdFee) GasPrices() sdk.DecCoins {
-	return sdk.NewDecCoinsFromCoins(fee.Amount...).QuoDec(sdk.NewDec(int64(fee.Gas)))
-}
-
-// Deprecated
-func NewStdSignature(pk crypto.PubKey, sig []byte) StdSignature {
-	return StdSignature{PubKey: pk, Signature: sig}
-}
-
-// GetSignature returns the raw signature bytes.
-func (ss StdSignature) GetSignature() []byte {
-	return ss.Signature
-}
-
-// GetPubKey returns the public key of a signature as a crypto.PubKey using the
-// Amino codec.
-func (ss StdSignature) GetPubKey() crypto.PubKey {
-	return ss.PubKey
-}
-
-// MarshalYAML returns the YAML representation of the signature.
-func (ss StdSignature) MarshalYAML() (interface{}, error) {
-	var (
-		bz     []byte
-		pubkey string
-		err    error
-	)
-
-	if ss.PubKey != nil {
-		pubkey, err = sdk.Bech32ifyPubKey(sdk.Bech32PubKeyTypeAccPub, ss.GetPubKey())
-		if err != nil {
-			return nil, err
-		}
-	}
-
-	bz, err = yaml.Marshal(struct {
-		PubKey    string
-		Signature string
-	}{
-		PubKey:    pubkey,
-		Signature: fmt.Sprintf("%X", ss.Signature),
-	})
-	if err != nil {
-		return nil, err
-	}
-
-	return string(bz), err
-}
-
-// CountSubKeys counts the total number of keys for a multi-sig public key.
-func CountSubKeys(pub crypto.PubKey) int {
-	v, ok := pub.(multisig.PubKeyMultisigThreshold)
-	if !ok {
-		return 1
-	}
-
-	numKeys := 0
-	for _, subkey := range v.PubKeys {
-		numKeys += CountSubKeys(subkey)
-	}
-
-	return numKeys
-}
-
-// ---------------------------------------------------------------------------
-// DEPRECATED
-// ---------------------------------------------------------------------------
-
-var (
-	_ sdk.Tx             = (*StdTx)(nil)
-	_ codectypes.IntoAny = (*StdTx)(nil)
-)
-
-// StdTx is the legacy transaction format for wrapping a Msg with Fee and Signatures.
-// It only works with Amino, please prefer the new protobuf Tx in types/tx.
-// NOTE: the first signature is the fee payer (Signatures must not be nil).
-type StdTx struct {
-	Msgs          []sdk.Msg      `json:"msg" yaml:"msg"`
-	Fee           StdFee         `json:"fee" yaml:"fee"`
-	Signatures    []StdSignature `json:"signatures" yaml:"signatures"`
-	Memo          string         `json:"memo" yaml:"memo"`
-	TimeoutHeight uint64         `json:"timeout_height" yaml:"timeout_height"`
-}
-
-// Deprecated
-func NewStdTx(msgs []sdk.Msg, fee StdFee, sigs []StdSignature, memo string) StdTx {
-	return StdTx{
-		Msgs:       msgs,
-		Fee:        fee,
-		Signatures: sigs,
-		Memo:       memo,
-	}
-}
-
-// GetMsgs returns the all the transaction's messages.
-func (tx StdTx) GetMsgs() []sdk.Msg { return tx.Msgs }
-
-// ValidateBasic does a simple and lightweight validation check that doesn't
-// require access to any other information.
-func (tx StdTx) ValidateBasic() error {
-	stdSigs := tx.GetSignatures()
-
-	if tx.Fee.Gas > MaxGasWanted {
-		return sdkerrors.Wrapf(
-			sdkerrors.ErrInvalidRequest,
-			"invalid gas supplied; %d > %d", tx.Fee.Gas, MaxGasWanted,
-		)
-	}
-	if tx.Fee.Amount.IsAnyNegative() {
-		return sdkerrors.Wrapf(
-			sdkerrors.ErrInsufficientFee,
-			"invalid fee provided: %s", tx.Fee.Amount,
-		)
-	}
-	if len(stdSigs) == 0 {
-		return sdkerrors.ErrNoSignatures
-	}
-	if len(stdSigs) != len(tx.GetSigners()) {
-		return sdkerrors.Wrapf(
-			sdkerrors.ErrUnauthorized,
-			"wrong number of signers; expected %d, got %d", len(tx.GetSigners()), len(stdSigs),
-		)
-	}
-
-	return nil
-}
-
-// AsAny implements IntoAny.AsAny.
-func (tx *StdTx) AsAny() *codectypes.Any {
-	return codectypes.UnsafePackAny(tx)
-}
-
-// GetSigners returns the addresses that must sign the transaction.
-// Addresses are returned in a deterministic order.
-// They are accumulated from the GetSigners method for each Msg
-// in the order they appear in tx.GetMsgs().
-// Duplicate addresses will be omitted.
-func (tx StdTx) GetSigners() []sdk.AccAddress {
-	var signers []sdk.AccAddress
-	seen := map[string]bool{}
-
-	for _, msg := range tx.GetMsgs() {
-		for _, addr := range msg.GetSigners() {
-			if !seen[addr.String()] {
-				signers = append(signers, addr)
-				seen[addr.String()] = true
-			}
-		}
-	}
-
-	return signers
-}
-
-// GetMemo returns the memo
-func (tx StdTx) GetMemo() string { return tx.Memo }
-
-// GetTimeoutHeight returns the transaction's timeout height (if set).
-func (tx StdTx) GetTimeoutHeight() uint64 {
-	return tx.TimeoutHeight
-}
-
-// GetSignatures returns the signature of signers who signed the Msg.
-// CONTRACT: Length returned is same as length of
-// pubkeys returned from MsgKeySigners, and the order
-// matches.
-// CONTRACT: If the signature is missing (ie the Msg is
-// invalid), then the corresponding signature is
-// .Empty().
-func (tx StdTx) GetSignatures() [][]byte {
-	sigs := make([][]byte, len(tx.Signatures))
-	for i, stdSig := range tx.Signatures {
-		sigs[i] = stdSig.Signature
-	}
-	return sigs
-}
-
-// GetSignaturesV2 implements SigVerifiableTx.GetSignaturesV2
-func (tx StdTx) GetSignaturesV2() ([]signing.SignatureV2, error) {
-	res := make([]signing.SignatureV2, len(tx.Signatures))
-
-	for i, sig := range tx.Signatures {
-		var err error
-		res[i], err = StdSignatureToSignatureV2(legacy.Cdc, sig)
-		if err != nil {
-			return nil, sdkerrors.Wrapf(err, "Unable to convert signature %v to V2", sig)
-		}
-	}
-
-	return res, nil
-}
-
-// GetPubkeys returns the pubkeys of signers if the pubkey is included in the signature
-// If pubkey is not included in the signature, then nil is in the slice instead
-func (tx StdTx) GetPubKeys() []crypto.PubKey {
-	pks := make([]crypto.PubKey, len(tx.Signatures))
-
-	for i, stdSig := range tx.Signatures {
-		pks[i] = stdSig.GetPubKey()
-	}
-
-	return pks
-}
-
-// GetGas returns the Gas in StdFee
-func (tx StdTx) GetGas() uint64 { return tx.Fee.Gas }
-
-// GetFee returns the FeeAmount in StdFee
-func (tx StdTx) GetFee() sdk.Coins { return tx.Fee.Amount }
-
-// FeePayer returns the address that is responsible for paying fee
-// StdTx returns the first signer as the fee payer
-// If no signers for tx, return empty address
-func (tx StdTx) FeePayer() sdk.AccAddress {
-	if tx.GetSigners() != nil {
-		return tx.GetSigners()[0]
-	}
-	return sdk.AccAddress{}
-}
-
-// StdSignDoc is replay-prevention structure.
-// It includes the result of msg.GetSignBytes(),
-// as well as the ChainID (prevent cross chain replay)
-// and the Sequence numbers for each signature (prevent
-// inchain replay and enforce tx ordering per account).
-type StdSignDoc struct {
-	AccountNumber uint64            `json:"account_number" yaml:"account_number"`
-	Sequence      uint64            `json:"sequence" yaml:"sequence"`
-	TimeoutHeight uint64            `json:"timeout_height,omitempty" yaml:"timeout_height"`
-	ChainID       string            `json:"chain_id" yaml:"chain_id"`
-	Memo          string            `json:"memo" yaml:"memo"`
-	Fee           json.RawMessage   `json:"fee" yaml:"fee"`
-	Msgs          []json.RawMessage `json:"msgs" yaml:"msgs"`
-}
-
-// StdSignBytes returns the bytes to sign for a transaction.
-func StdSignBytes(chainID string, accnum, sequence, timeout uint64, fee StdFee, msgs []sdk.Msg, memo string) []byte {
-	msgsBytes := make([]json.RawMessage, 0, len(msgs))
-	for _, msg := range msgs {
-		msgsBytes = append(msgsBytes, json.RawMessage(msg.GetSignBytes()))
-	}
-
-	bz, err := legacy.Cdc.MarshalJSON(StdSignDoc{
-		AccountNumber: accnum,
-		ChainID:       chainID,
-		Fee:           json.RawMessage(fee.Bytes()),
-		Memo:          memo,
-		Msgs:          msgsBytes,
-		Sequence:      sequence,
-		TimeoutHeight: timeout,
-	})
-	if err != nil {
-		panic(err)
-	}
-
-	return sdk.MustSortJSON(bz)
-}
-
-// Deprecated: StdSignature represents a sig
-type StdSignature struct {
-	crypto.PubKey `json:"pub_key" yaml:"pub_key"` // optional
-	Signature     []byte                          `json:"signature" yaml:"signature"`
-}
-
-// DefaultTxDecoder logic for standard transaction decoding
-func DefaultTxDecoder(cdc *codec.LegacyAmino) sdk.TxDecoder {
-	return func(txBytes []byte) (sdk.Tx, error) {
-		var tx = StdTx{}
-
-		if len(txBytes) == 0 {
-			return nil, sdkerrors.Wrap(sdkerrors.ErrTxDecode, "tx bytes are empty")
-		}
-
-		// StdTx.Msg is an interface. The concrete types
-		// are registered by MakeTxCodec
-		err := cdc.UnmarshalBinaryBare(txBytes, &tx)
-		if err != nil {
-			return nil, sdkerrors.Wrap(sdkerrors.ErrTxDecode, err.Error())
-		}
-
-		return tx, nil
-	}
-}
-
-func DefaultJSONTxDecoder(cdc *codec.LegacyAmino) sdk.TxDecoder {
-	return func(txBytes []byte) (sdk.Tx, error) {
-		var tx = StdTx{}
-
-		if len(txBytes) == 0 {
-			return nil, sdkerrors.Wrap(sdkerrors.ErrTxDecode, "tx bytes are empty")
-		}
-
-		// StdTx.Msg is an interface. The concrete types
-		// are registered by MakeTxCodec
-		err := cdc.UnmarshalJSON(txBytes, &tx)
-		if err != nil {
-			return nil, sdkerrors.Wrap(sdkerrors.ErrTxDecode, err.Error())
-		}
-
-		return tx, nil
-	}
-}
-
-// DefaultTxEncoder logic for standard transaction encoding
-func DefaultTxEncoder(cdc *codec.LegacyAmino) sdk.TxEncoder {
-	return func(tx sdk.Tx) ([]byte, error) {
-		return cdc.MarshalBinaryBare(tx)
-	}
-}
-
-var _ codectypes.UnpackInterfacesMessage = StdTx{}
-
-func (tx StdTx) UnpackInterfaces(unpacker codectypes.AnyUnpacker) error {
-	for _, m := range tx.Msgs {
-		err := codectypes.UnpackInterfaces(m, unpacker)
-		if err != nil {
-			return err
-		}
-	}
-	return nil
-}
-
-// StdSignatureToSignatureV2 converts a StdSignature to a SignatureV2
-func StdSignatureToSignatureV2(cdc *codec.LegacyAmino, sig StdSignature) (signing.SignatureV2, error) {
-	pk := sig.GetPubKey()
-	data, err := pubKeySigToSigData(cdc, pk, sig.Signature)
-	if err != nil {
-		return signing.SignatureV2{}, err
-	}
-
-	return signing.SignatureV2{
-		PubKey:            pk,
-		Data:              data,
-		SkipSequenceCheck: true,
-	}, nil
-}
-
-// SignatureV2ToStdSignature converts a SignatureV2 to a StdSignature
-func SignatureV2ToStdSignature(cdc *codec.LegacyAmino, sig signing.SignatureV2) (StdSignature, error) {
-	var (
-		sigBz []byte
-		err   error
-	)
-
-	if sig.Data != nil {
-		sigBz, err = SignatureDataToAminoSignature(cdc, sig.Data)
-		if err != nil {
-			return StdSignature{}, err
-		}
-	}
-
-	return StdSignature{
-		PubKey:    sig.PubKey,
-		Signature: sigBz,
-	}, nil
-}
-
-func pubKeySigToSigData(cdc *codec.LegacyAmino, key crypto.PubKey, sig []byte) (signing.SignatureData, error) {
-	multiPK, ok := key.(multisig.PubKey)
-	if !ok {
-		return &signing.SingleSignatureData{
-			SignMode:  signing.SignMode_SIGN_MODE_LEGACY_AMINO_JSON,
-			Signature: sig,
-		}, nil
-	}
-	var multiSig multisig.AminoMultisignature
-	err := cdc.UnmarshalBinaryBare(sig, &multiSig)
-	if err != nil {
-		return nil, err
-	}
-
-	sigs := multiSig.Sigs
-	sigDatas := make([]signing.SignatureData, len(sigs))
-	pubKeys := multiPK.GetPubKeys()
-	bitArray := multiSig.BitArray
-	n := multiSig.BitArray.Count()
-	signatures := multisig.NewMultisig(n)
-	sigIdx := 0
-	for i := 0; i < n; i++ {
-		if bitArray.GetIndex(i) {
-			data, err := pubKeySigToSigData(cdc, pubKeys[i], multiSig.Sigs[sigIdx])
-			if err != nil {
-				return nil, sdkerrors.Wrapf(err, "Unable to convert Signature to SigData %d", sigIdx)
-			}
-
-			sigDatas[sigIdx] = data
-			multisig.AddSignature(signatures, data, sigIdx)
-			sigIdx++
-		}
-	}
-
-	return signatures, nil
-}
-
-// MultiSignatureDataToAminoMultisignature converts a MultiSignatureData to an AminoMultisignature.
-// Only SIGN_MODE_LEGACY_AMINO_JSON is supported.
-func MultiSignatureDataToAminoMultisignature(cdc *codec.LegacyAmino, mSig *signing.MultiSignatureData) (multisig.AminoMultisignature, error) {
-	n := len(mSig.Signatures)
-	sigs := make([][]byte, n)
-
-	for i := 0; i < n; i++ {
-		var err error
-		sigs[i], err = SignatureDataToAminoSignature(cdc, mSig.Signatures[i])
-		if err != nil {
-			return multisig.AminoMultisignature{}, sdkerrors.Wrapf(err, "Unable to convert Signature Data to signature %d", i)
-		}
-	}
-
-	return multisig.AminoMultisignature{
-		BitArray: mSig.BitArray,
-		Sigs:     sigs,
-	}, nil
-}
-
-// SignatureDataToAminoSignature converts a SignatureData to amino-encoded signature bytes.
-// Only SIGN_MODE_LEGACY_AMINO_JSON is supported.
-func SignatureDataToAminoSignature(cdc *codec.LegacyAmino, data signing.SignatureData) ([]byte, error) {
-	switch data := data.(type) {
-	case *signing.SingleSignatureData:
-		if data.SignMode != signing.SignMode_SIGN_MODE_LEGACY_AMINO_JSON {
-			return nil, fmt.Errorf("expected %s, got %s", signing.SignMode_SIGN_MODE_LEGACY_AMINO_JSON, data.SignMode)
-		}
-
-		return data.Signature, nil
-	case *signing.MultiSignatureData:
-		aminoMSig, err := MultiSignatureDataToAminoMultisignature(cdc, data)
-		if err != nil {
-			return nil, err
-		}
-
-		return cdc.MarshalBinaryBare(aminoMSig)
-	default:
-		return nil, fmt.Errorf("unexpected signature data %T", data)
-	}
-}
