+++ x/bank/client/rest/tx_test.go
@@ -11,6 +11,7 @@ import (
 	"github.com/cosmos/cosmos-sdk/types/errors"
 	"github.com/cosmos/cosmos-sdk/types/rest"
 	authclient "github.com/cosmos/cosmos-sdk/x/auth/client"
+	"github.com/cosmos/cosmos-sdk/x/auth/legacy/legacytx"
 	authtypes "github.com/cosmos/cosmos-sdk/x/auth/types"
 	bankrest "github.com/cosmos/cosmos-sdk/x/bank/client/rest"
 	banktypes "github.com/cosmos/cosmos-sdk/x/bank/types"
@@ -43,25 +44,25 @@ func (s *IntegrationTestSuite) TestCoinSend() {
 	}, stdTx.GetMsgs())
 }
 
-func submitSendReq(val *network.Validator, req bankrest.SendReq) (authtypes.StdTx, error) {
+func submitSendReq(val *network.Validator, req bankrest.SendReq) (legacytx.StdTx, error) {
 	url := fmt.Sprintf("%s/bank/accounts/%s/transfers", val.APIAddress, val.Address)
 
 	// NOTE: this uses amino explicitly, don't migrate it!
 	bz, err := val.ClientCtx.LegacyAmino.MarshalJSON(req)
 	if err != nil {
-		return authtypes.StdTx{}, errors.Wrap(err, "error encoding SendReq to json")
+		return legacytx.StdTx{}, errors.Wrap(err, "error encoding SendReq to json")
 	}
 
 	res, err := rest.PostRequest(url, "application/json", bz)
 	if err != nil {
-		return authtypes.StdTx{}, err
+		return legacytx.StdTx{}, err
 	}
 
-	var tx authtypes.StdTx
+	var tx legacytx.StdTx
 	// NOTE: this uses amino explicitly, don't migrate it!
 	err = val.ClientCtx.LegacyAmino.UnmarshalJSON(res, &tx)
 	if err != nil {
-		return authtypes.StdTx{}, errors.Wrap(err, "error unmarshaling to StdTx SendReq response")
+		return legacytx.StdTx{}, errors.Wrap(err, "error unmarshaling to StdTx SendReq response")
 	}
 
 	return tx, nil
